move edit menu item to text content fragment
