+++ app/src/main/java/org/stepik/android/view/lesson/ui/activity/LessonActivity.kt
@@ -149,6 +149,7 @@ class LessonActivity : FragmentActivityBase(), LessonView,
             override fun onPageSelected(position: Int) {
                 currentFocus?.hideKeyboard()
                 lessonPresenter.onStepOpened(position)
+                invalidateOptionsMenu()
             }
         })
         lessonTab.setupWithViewPager(lessonPager, true)
