+++ app/src/main/res/menu/step_edit_menu.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/menu_item_edit"
+        android:icon="@drawable/ic_edit"
+        android:title="@string/step_action_edit_content"
+        app:showAsAction="ifRoom" />
+</menu>
